+++ server/graphql/index.ts
@@ -1,0 +1,9 @@
+import { GraphQLSchema } from 'graphql';
+
+import RootQueryType from './query';
+
+const schema = new GraphQLSchema({
+  query: RootQueryType,
+});
+
+export default schema;
