+++ server/routes/posts/tests.spec.ts
@@ -2,7 +2,7 @@ import http from 'http';
 import request from 'supertest';
 import setCookie from 'set-cookie-parser';
 
-import server from '../server';
+import server from '../../server';
 
 interface iUser {
   _id?: string;
