+++ server/routes/users/routes.ts
@@ -7,8 +7,8 @@ import {
   retrieveUserByGoogleId,
   toggleFollow,
   updateUser,
-} from '../controllers/user';
-import auth from '../middlewares/authenticator';
+} from './controllers';
+import auth from '../../middlewares/authenticator';
 
 const router = express.Router();
